api/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via jsonschema
boto3==1.21.35
    # via -r requirements.in
botocore==1.24.35
    # via
    #   boto3
    #   s3transfer
brotlipy==0.7.0
    # via -r requirements.in
cffi==1.15.0
    # via brotlipy
defusedxml==0.7.1
    # via odfpy
diff-match-patch==20200713
    # via django-import-export
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.12
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-import-export
    #   django-jquery
    #   django-modeltranslation
    #   django-storages
    #   django-taggit
    #   djangorestframework
    #   drf-spectacular
django-ckeditor==6.3.0
    # via -r requirements.in
django-cors-headers==3.11.0
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-import-export==2.8.0
    # via -r requirements.in
django-jquery==3.1.0
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-ckeditor
django-modeltranslation==0.17.5
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
django-taggit==2.1.0
    # via
    #   -r requirements.in
    #   django-taggit-serializer
django-taggit-serializer==0.1.7
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.22.0
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
gevent==21.12.0
    # via -r requirements.in
greenlet==1.1.2
    # via gevent
gunicorn==20.1.0
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonschema==4.4.0
    # via drf-spectacular
markuppy==1.14
    # via tablib
odfpy==1.4.1
    # via tablib
openpyxl==3.0.9
    # via tablib
pillow==9.1.0
    # via -r requirements.in
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.21
    # via cffi
pymemcache==3.5.2
    # via -r requirements.in
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via botocore
pytz==2021.3
    # via
    #   django
    #   djangorestframework
pyyaml==6.0
    # via
    #   drf-spectacular
    #   tablib
s3transfer==0.5.1
    # via boto3
six==1.16.0
    # via
    #   django-modeltranslation
    #   django-taggit-serializer
    #   pymemcache
    #   python-dateutil
sqlparse==0.4.2
    # via django
tablib[html,ods,xls,xlsx,yaml]==3.2.0
    # via
    #   -r requirements.in
    #   django-import-export
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.8
    # via botocore
whitenoise==6.0.0
    # via -r requirements.in
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
zope-event==4.5.0
    # via gevent
zope-interface==5.4.0
    # via gevent

# The following packages are considered to be unsafe in a requirements file:
# setuptools
